added authbuttons from navbar to landing pg
